Move handling binary responses to a shared space
